+++ src/ansys/mechanical/core/embedding/app.py
@@ -1,10 +1,10 @@
 """Main application class for embedded Mechanical."""
 import atexit
 import os
-import shutil
 
 from ansys.mechanical.core.embedding import initializer, runtime
 from ansys.mechanical.core.embedding.addins import AddinConfiguration
+from ansys.mechanical.core.embedding.appdata import UniqueUserProfile
 
 
 def _get_default_addin_configuration() -> AddinConfiguration:
@@ -21,8 +21,8 @@ def _dispose_embedded_app(instances):  # pragma: nocover
         instance._dispose()
 
 
-def _cleanup_private_appdata(folder):
-    shutil.rmtree(folder, ignore_errors=True)
+def _cleanup_private_appdata(profile: UniqueUserProfile):
+    profile.cleanup()
 
 
 def _start_application(configuration: AddinConfiguration, version, db_file) -> "App":
@@ -74,9 +74,11 @@ class App:
         configuration = kwargs.get("config", _get_default_addin_configuration())
 
         if private_appdata:
-            self.pid = os.getpid()
-            self.tmp_dir = initializer.set_private_appdata(self.pid)
-            atexit.register(_cleanup_private_appdata, self.tmp_dir)
+            new_profile_name = f"PyMechanical-{os.getpid()}"
+            profile = UniqueUserProfile(new_profile_name)
+            profile.warn()
+            profile.update_environment(os.environ)
+            atexit.register(_cleanup_private_appdata, profile)
 
         self._app = _start_application(configuration, self._version, db_file)
         runtime.initialize()